requirements.txt

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.4.0
django-filter==23.4
django-extensions==3.2.3
asgiref==3.9.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2

# Autenticación y gestión de usuarios
django-allauth==0.51.0

# Tareas asíncronas
celery==5.2.7

# Mejora en la entrega de archivos estáticos
whitenoise==5.3.0

# Testing
pytest-django==4.5.2

# Docs
drf-spectacular==0.24.0
